1-added edit function, 2-limited texarea text to 50k chars, 3-limited post lines to 10
