+++ src/components/userPosts/UserPosts.js
@@ -6,7 +6,7 @@ import { getUserPosts } from '../../services/userPostsApi.js';
 import { getUserInfo } from '../../services/userInfoApi.js';
 import NavBar from "../navBar/NavBar";
 import Trending from "../shared/Trending";
-import Post from '../shared/Post.js';
+import Post from '../shared/post/Post.js';
 import Loading from '../shared/Loading.js';
 
 export default function MyPosts () {
