+++ src/styles/globalStyles.js
@@ -5,6 +5,10 @@ const GlobalStyle = createGlobalStyle`
     background-color: #333333;
     font-family: 'Lato', sans-serif;
   }
+
+  * {
+      word-wrap: break-word;
+  }
 `
 
 export default GlobalStyle;
